requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-socketio==5.3.6
pymysql==1.1.0
lightgbm==4.0.0
pandas==2.1.1
numpy==1.26.4
scipy==1.11.3
scikit-learn==1.3.2
joblib==1.3.2
opencv-python==4.8.1.78
Pillow==10.0.1
websockets==11.0.3
python-socketio==5.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
matplotlib==3.8.4
seaborn==0.12.2
plotly==5.17.0
requests==2.31.0
qrcode==7.4.2


